execution-agent/requirements.txt
fastapi==0.95.2
pydantic==1.10.14
uvicorn[standard]==0.29.0
ansible-core==2.16.6
python-dotenv==1.0.0